% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{centers}
\alias{centers}
\title{Cluster centers}
\usage{
centers(x, ...)
}
\arguments{
\item{x}{A clustering object.}

\item{...}{Unused.}
}
\value{
A matrix or \code{NULL}.
}
\description{
Cluster centers
}
