% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{clusters}
\alias{clusters}
\title{Cluster assignments}
\usage{
clusters(x, ...)
}
\arguments{
\item{x}{A clustering object.}

\item{...}{Unused.}
}
\value{
An integer vector.
}
\description{
Cluster assignments
}
