% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_internal.R
\name{compute_calinski_harabasz}
\alias{compute_calinski_harabasz}
\title{Calinski-Harabasz index}
\usage{
compute_calinski_harabasz(data, clusters)
}
\description{
The Calinski-Harabasz index is
}
\details{
\deqn{
\mathrm{CH} = \frac{\mathrm{BSS}/(k - 1)}{\mathrm{WSS}/(n - k)},
}

where \eqn{k} is the number of clusters, \eqn{n} is the number of
observations, \eqn{\mathrm{BSS}} is the between-cluster sum of squares, and
\eqn{\mathrm{WSS}} is the within-cluster sum of squares.
}
