% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_internal.R
\name{compute_davies_bouldin}
\alias{compute_davies_bouldin}
\title{Davies-Bouldin index}
\usage{
compute_davies_bouldin(data, clusters)
}
\description{
The Davies-Bouldin index is
}
\details{
\deqn{
\mathrm{DB} = \frac{1}{k} \sum_{j=1}^k \max_{l \ne j}
\frac{s_j + s_l}{d(\mu_j, \mu_l)},
}

where \eqn{s_j} is the average scatter of cluster \eqn{j},
\eqn{\mu_j} is its centroid, and \eqn{d(\mu_j, \mu_l)} is the distance
between cluster centroids.
}
