% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_internal.R
\name{compute_silhouette_metric}
\alias{compute_silhouette_metric}
\title{Mean silhouette width}
\usage{
compute_silhouette_metric(data, clusters)
}
\description{
The silhouette width of an observation is
}
\details{
\deqn{
s(i) = \frac{b(i) - a(i)}{\max\{a(i), b(i)\}},
}

where \eqn{a(i)} is the mean distance from observation \eqn{i} to the other
observations in its own cluster and \eqn{b(i)} is the minimum mean distance
from \eqn{i} to the observations in any other cluster. The returned value is
the mean over all observations.
}
