% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_internal.R
\name{compute_total_within}
\alias{compute_total_within}
\title{Total within-cluster sum of squares}
\usage{
compute_total_within(data, clusters)
}
\description{
The total within-cluster dispersion is
}
\details{
\deqn{
\mathrm{WSS} = \sum_{j=1}^k \sum_{i \in C_j} \|x_i - \mu_j\|^2.
}
}
