% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_kmm.R
\name{lambda_kmm}
\alias{lambda_kmm}
\title{Estimate the KMM lambda weight}
\usage{
lambda_kmm(data, scale_numeric = TRUE)
}
\arguments{
\item{data}{Mixed-type data.}

\item{scale_numeric}{Logical.}
}
\description{
Estimate the KMM lambda weight
}
\keyword{internal}
