% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{membership}
\alias{membership}
\title{Membership matrix}
\usage{
membership(x, ...)
}
\arguments{
\item{x}{A clustering object.}

\item{...}{Unused.}
}
\value{
A matrix or \code{NULL}.
}
\description{
Membership matrix
}
