% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{method_used}
\alias{method_used}
\title{Method used}
\usage{
method_used(x, ...)
}
\arguments{
\item{x}{A clustering object.}

\item{...}{Unused.}
}
\value{
Character scalar.
}
\description{
Method used
}
