% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{n_clusters}
\alias{n_clusters}
\title{Number of clusters}
\usage{
n_clusters(x, ...)
}
\arguments{
\item{x}{A clustering object.}

\item{...}{Unused.}
}
\value{
Integer scalar.
}
\description{
Number of clusters
}
