% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_external.R
\name{normalized_mutual_information}
\alias{normalized_mutual_information}
\title{Normalized mutual information}
\usage{
normalized_mutual_information(x, y)
}
\description{
\deqn{
\mathrm{NMI}(x, y) = \frac{I(x;y)}{\sqrt{H(x)H(y)}},
}
}
\details{
where \eqn{I} is mutual information and \eqn{H} is Shannon entropy.
}
