% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_clusters.R
\name{plot_cluster_sizes}
\alias{plot_cluster_sizes}
\title{Plot cluster sizes}
\usage{
plot_cluster_sizes(x, ...)
}
\arguments{
\item{x}{A \code{cluster_fit} or \code{cluster_explore} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot cluster sizes
}
