% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_metacluster.R
\name{plot_coassoc}
\alias{plot_coassoc}
\title{Plot the co-association matrix}
\usage{
plot_coassoc(x, ...)
}
\arguments{
\item{x}{A \code{metacluster_fit} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot the co-association matrix
}
