% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_dendrogram.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot a clustering dendrogram}
\usage{
plot_dendrogram(x, ...)
}
\arguments{
\item{x}{A hierarchical \code{cluster_fit} or \code{metacluster_fit} object.}

\item{...}{Unused.}
}
\value{
The underlying fitted dendrogram object, invisibly.
}
\description{
Plot a clustering dendrogram
}
