% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_clusters.R
\name{plot_feature_profiles}
\alias{plot_feature_profiles}
\title{Plot feature profiles by cluster}
\usage{
plot_feature_profiles(x, features = NULL, ...)
}
\arguments{
\item{x}{A \code{cluster_explore} object.}

\item{features}{Optional character vector of features to plot.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot feature profiles by cluster
}
