% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_silhouette}
\alias{plot_silhouette}
\title{Plot silhouette widths}
\usage{
plot_silhouette(x, ...)
}
\arguments{
\item{x}{A \code{cluster_fit} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot silhouette widths
}
