% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_validation.R
\name{plot_validation}
\alias{plot_validation}
\title{Plot validation metrics}
\usage{
plot_validation(x, ...)
}
\arguments{
\item{x}{A \code{cluster_validation} object.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot validation metrics
}
