% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_kmm.R
\name{predict_kmm}
\alias{predict_kmm}
\title{Predict with KMM}
\usage{
predict_kmm(object, new_data, ...)
}
\arguments{
\item{object}{A \code{cluster_fit} object.}

\item{new_data}{Mixed-type data frame.}
}
\description{
Predict with KMM
}
\keyword{internal}
