% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{prototypes}
\alias{prototypes}
\title{Cluster prototypes}
\usage{
prototypes(x, ...)
}
\arguments{
\item{x}{A clustering object.}

\item{...}{Unused.}
}
\value{
A matrix, data frame, or \code{NULL}.
}
\description{
Cluster prototypes
}
