% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_kmm.R
\name{validate_kmm_params}
\alias{validate_kmm_params}
\title{Validation helpers for KMM}
\usage{
validate_kmm_params(data, params)
}
\arguments{
\item{data}{A data frame or matrix.}

\item{params}{Method parameters.}
}
\description{
Validation helpers for KMM
}
\keyword{internal}
