% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_internal.R
\name{compute_per_cluster_silhouette}
\alias{compute_per_cluster_silhouette}
\title{Mean silhouette width by cluster}
\usage{
compute_per_cluster_silhouette(data, clusters)
}
\description{
This returns the average silhouette width within each cluster.
}
