% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{method_used}
\alias{method_used}
\title{Method used}
\usage{
method_used(x, ...)
}
\arguments{
\item{x}{A \code{cluster_fit} or \code{metacluster_fit} object.}

\item{...}{Unused.}
}
\value{
Character scalar (e.g. \code{"kmeans"}, \code{"pam"}, \code{"hclust"}).
}
\description{
Return the name of the clustering method used to fit the object.
}
\examples{
fit <- cluster(iris[, 1:4], method = "kmeans", k = 3, seed = 1)
method_used(fit)
}
\seealso{
\code{\link[=cluster]{cluster()}}
}
