% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_clusters.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clustered observations in 2D}
\usage{
plot_clusters(x, data = NULL, ...)
}
\arguments{
\item{x}{A \code{cluster_fit} or \code{cluster_explore} object.}

\item{data}{Optional numeric data used to compute the embedding.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot clustered observations in 2D
}
