% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_metacluster.R
\name{plot_coassoc}
\alias{plot_coassoc}
\title{Plot the co-association matrix}
\usage{
plot_coassoc(x, ...)
}
\arguments{
\item{x}{A \code{metacluster_fit} object.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Display the \eqn{n \times n} co-association matrix as a heatmap. High values
(near 1) indicate pairs of observations that were consistently co-clustered
across candidate partitions.
}
\examples{
mfit <- metacluster(iris[, 1:4], methods = c("kmeans", "hclust"),
                    k = 2:3, seed = 1)
plot_coassoc(mfit)
}
\seealso{
\code{\link[=metacluster]{metacluster()}}, \code{\link[=plot_consensus]{plot_consensus()}}
}
