% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_metacluster.R
\name{plot_consensus}
\alias{plot_consensus}
\title{Plot consensus clustering in 2D}
\usage{
plot_consensus(x, data = NULL, ...)
}
\arguments{
\item{x}{A \code{metacluster_fit} object.}

\item{data}{Optional numeric matrix or data frame used to compute the
embedding. Defaults to the training data stored in \code{x}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Project the consensus cluster assignments onto a two-dimensional PCA
embedding.
}
\examples{
mfit <- metacluster(iris[, 1:4], methods = c("kmeans", "hclust"),
                    k = 2:3, seed = 1)
plot_consensus(mfit)
}
\seealso{
\code{\link[=metacluster]{metacluster()}}, \code{\link[=plot_coassoc]{plot_coassoc()}}
}
