% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_metacluster.R
\name{plot_consensus}
\alias{plot_consensus}
\title{Plot consensus clustering in 2D}
\usage{
plot_consensus(x, data = NULL, ...)
}
\arguments{
\item{x}{A \code{metacluster_fit} object.}

\item{data}{Optional numeric data used to compute the embedding.}

\item{...}{Unused.}
}
\value{
A ggplot object.
}
\description{
Plot consensus clustering in 2D
}
